Refactors `types::r#type::IonType` to `types::IonType`.
This addresses a papercut where some IDE integration gets confused with a
module that requires a raw identifier to express it, but also eliminates
a "stuttering" for `types::type` which had only the `IonType in it and
an aliasing back to `types/mod.rs` anyhow that I found a bit confusing
at times.
